+++ tpu/src/test/java/tpu/TpuVmIT.java
@@ -17,6 +17,7 @@
 package tpu;
 
 import static com.google.common.truth.Truth.assertThat;
+import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
@@ -25,14 +26,18 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import com.google.api.gax.longrunning.OperationFuture;
+import com.google.cloud.tpu.v2.CreateNodeRequest;
 import com.google.cloud.tpu.v2.DeleteNodeRequest;
 import com.google.cloud.tpu.v2.GetNodeRequest;
+import com.google.cloud.tpu.v2.ListNodesRequest;
 import com.google.cloud.tpu.v2.Node;
 import com.google.cloud.tpu.v2.TpuClient;
 import com.google.cloud.tpu.v2.TpuSettings;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
 import java.io.PrintStream;
+import java.util.Arrays;
+import java.util.List;
 import java.util.concurrent.ExecutionException;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -47,6 +52,8 @@ public class TpuVmIT {
   private static final String PROJECT_ID = "project-id";
   private static final String ZONE = "asia-east1-c";
   private static final String NODE_NAME = "test-tpu";
+  private static final String TPU_TYPE = "v2-8";
+  private static final String TPU_SOFTWARE_VERSION = "tpu-vm-tf-2.12.1";
   private static ByteArrayOutputStream bout;
 
   @BeforeAll
@@ -56,20 +63,44 @@ public class TpuVmIT {
   }
 
   @Test
+  public void testCreateTpuVm() throws Exception {
+    try (MockedStatic<TpuClient> mockedTpuClient = mockStatic(TpuClient.class)) {
+      Node mockNode = mock(Node.class);
+      TpuClient mockTpuClient = mock(TpuClient.class);
+      OperationFuture mockFuture = mock(OperationFuture.class);
+
+      mockedTpuClient.when(() -> TpuClient.create(any(TpuSettings.class)))
+          .thenReturn(mockTpuClient);
+      when(mockTpuClient.createNodeAsync(any(CreateNodeRequest.class)))
+          .thenReturn(mockFuture);
+      when(mockFuture.get()).thenReturn(mockNode);
+
+      Node returnedNode = CreateTpuVm.createTpuVm(
+          PROJECT_ID, ZONE, NODE_NAME,
+          TPU_TYPE, TPU_SOFTWARE_VERSION);
+
+      verify(mockTpuClient, times(1))
+          .createNodeAsync(any(CreateNodeRequest.class));
+      verify(mockFuture, times(1)).get();
+      assertEquals(returnedNode, mockNode);
+    }
+  }
+
+  @Test
   public void testGetTpuVm() throws IOException {
     try (MockedStatic<TpuClient> mockedTpuClient = mockStatic(TpuClient.class)) {
       Node mockNode = mock(Node.class);
       TpuClient mockClient = mock(TpuClient.class);
-      GetTpuVm mockGetTpuVm = mock(GetTpuVm.class);
 
       mockedTpuClient.when(TpuClient::create).thenReturn(mockClient);
       when(mockClient.getNode(any(GetNodeRequest.class))).thenReturn(mockNode);
 
       Node returnedNode = GetTpuVm.getTpuVm(PROJECT_ID, ZONE, NODE_NAME);
 
-      verify(mockGetTpuVm, times(1))
-          .getTpuVm(PROJECT_ID, ZONE, NODE_NAME);
+      verify(mockClient, times(1))
+          .getNode(any(GetNodeRequest.class));
       assertThat(returnedNode).isEqualTo(mockNode);
+      verify(mockClient, times(1)).close();
     }
   }
 
@@ -89,6 +120,29 @@ public class TpuVmIT {
 
       assertThat(output).contains("TPU VM deleted");
       verify(mockTpuClient, times(1)).deleteNodeAsync(any(DeleteNodeRequest.class));
+    }
+  }
+
+  @Test
+  public void testListTpuVm() throws IOException {
+    try (MockedStatic<TpuClient> mockedTpuClient = mockStatic(TpuClient.class)) {
+      Node mockNode1 = mock(Node.class);
+      Node mockNode2 = mock(Node.class);
+      List<Node> mockListNodes = Arrays.asList(mockNode1, mockNode2);
+
+      TpuClient mockTpuClient = mock(TpuClient.class);
+      mockedTpuClient.when(TpuClient::create).thenReturn(mockTpuClient);
+      TpuClient.ListNodesPagedResponse mockListNodesResponse =
+          mock(TpuClient.ListNodesPagedResponse.class);
+      when(mockTpuClient.listNodes(any(ListNodesRequest.class))).thenReturn(mockListNodesResponse);
+      TpuClient.ListNodesPage mockListNodesPage = mock(TpuClient.ListNodesPage.class);
+      when(mockListNodesResponse.getPage()).thenReturn(mockListNodesPage);
+      when(mockListNodesPage.getValues()).thenReturn(mockListNodes);
+
+      TpuClient.ListNodesPage returnedListNodes = ListTpuVms.listTpuVms(PROJECT_ID, ZONE);
+
+      assertThat(returnedListNodes.getValues()).isEqualTo(mockListNodes);
+      verify(mockTpuClient, times(1)).listNodes(any(ListNodesRequest.class));
     }
   }
 }
